refactor: user API 서버/클라이언트 파일 분리 (#79)
