+++ api/user.api.ts
@@ -1,27 +1,9 @@
-import { cache } from 'react';
-import { cookies } from 'next/headers';
+import apiClient from './axios';
 import { User } from '@/types/user.type';
 
-export const getProfile = cache(async (): Promise<User | null> => {
-  const cookieStore = await cookies();
-  const accessToken = cookieStore.get('accessToken')?.value;
-  if (!accessToken) return null;
-
-  try {
-    const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/v1/users/profile`, {
-      headers: { Authorization: `Bearer ${accessToken}` },
-    });
-
-    if (response.status === 401) {
-      cookieStore.delete('accessToken');
-      return null;
-    }
-
-    if (!response.ok) return null;
-
-    const data = await response.json();
-    return data.data;
-  } catch {
-    return null;
-  }
-});
+export const updateProfile = async (nickname: string): Promise<User> => {
+  const response = await apiClient.patch('/api/v1/users/profile', nickname, {
+    headers: { 'Content-Type': 'application/json' },
+  });
+  return response.data.data;
+};
